requirements.txt
langchain==0.1.20
langchain-groq==0.1.5
langchain-community==0.1.30
langchain-huggingface==0.0.16
langchain-chroma==0.1.2
langgraph==0.1.37
chromadb==0.4.24
python-dotenv==1.0.0
sentence-transformers==2.2.2
pydantic==2.5.3
groq==0.4.2
streamlit==1.28.0
